worker/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
brotlipy==0.7.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.1
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
pika==1.3.0
pip==22.1.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
setuptools==63.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
zope.event==4.5.0
zope.interface==5.4.0
